+++ packages/handlersjs-core/lib/daemons/scheduler.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable jest/expect-expect */
 import { lastValueFrom, Observable, of } from 'rxjs';
 import { Handler } from '../handlers/handler';
 import { Scheduler } from './scheduler';
@@ -8,7 +9,7 @@ describe('Scheduler', () => {
   const task = jest.fn();
   class TestHandler implements Handler<void, void> {
 
-    handle(input: void, intermediateOutput?: void): Observable<void> {
+    handle(): Observable<void> {
 
       task();
 
@@ -18,18 +19,20 @@ describe('Scheduler', () => {
 
   }
 
-  const expectIsRunning = async () => {
+  const expectIsRunning = () => {
 
     task.mockClear();
     jest.advanceTimersByTime(20);
-    expect(task).toBeCalledTimes(4);
+
+    expect(task).toHaveBeenCalledTimes(4);
 
   };
 
-  const expectIsNotRunning = async () => {
+  const expectIsNotRunning = () => {
 
     task.mockClear();
     jest.advanceTimersByTime(20);
+
     expect(task).not.toHaveBeenCalled();
 
   };
@@ -45,13 +48,13 @@ describe('Scheduler', () => {
   afterEach(async () => {
 
     // always stop the scheduler, ignore if an error "Scheduler wasn't running" is thrown
-    await lastValueFrom(scheduler.stop()).catch((e) => undefined);
+    await lastValueFrom(scheduler.stop()).catch(() => undefined);
 
   });
 
-  it('does not run the task on initialize', async () => {
+  it('does not run the task on initialize', () => {
 
-    await expectIsNotRunning();
+    expectIsNotRunning();
 
   });
 
@@ -60,13 +63,14 @@ describe('Scheduler', () => {
     it('will schedule the task when started', async () => {
 
       await lastValueFrom(scheduler.start());
-      await expectIsRunning();
+      expectIsRunning();
 
     });
 
     it('can not start when it was already running', async() => {
 
       await lastValueFrom(scheduler.start());
+
       await expect(lastValueFrom(scheduler.start())).rejects.toThrow('Scheduler was already running');
 
     });
@@ -80,7 +84,7 @@ describe('Scheduler', () => {
       await lastValueFrom(scheduler.start());
       await lastValueFrom(scheduler.stop());
 
-      await expectIsNotRunning();
+      expectIsNotRunning();
 
     });
 
@@ -90,7 +94,7 @@ describe('Scheduler', () => {
       await lastValueFrom(scheduler.stop());
       await lastValueFrom(scheduler.start());
 
-      await expectIsRunning();
+      expectIsRunning();
 
     });