+++ packages/handlersjs-core/lib/handlers/typed-pipe-through.handler.ts
@@ -1,4 +1,4 @@
-import { Observable, of, throwError } from 'rxjs';
+import { Observable } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { HandlerArgumentError } from '../errors/handler-argument-error';
 import { Handler } from './handler';