+++ packages/handlersjs-http/lib/errors/http-handler-error.spec.ts
@@ -1,3 +1,4 @@
+import { HttpHandlerResponse } from '../models/http-handler-response';
 import { HttpHandlerError } from './http-handler-error';
 
 describe('HttpHandlerRoute', () => {
@@ -6,7 +7,8 @@ describe('HttpHandlerRoute', () => {
 
     it('should be instantiated correctly', () => {
 
-      const error = new HttpHandlerError('label', undefined, undefined);
+      const error = new HttpHandlerError('label', (undefined as unknown as number), (undefined as unknown as HttpHandlerResponse));
+
       expect(error).toBeDefined();
       expect(error).toBeTruthy();