+++ packages/handlersjs-http/lib/handlers/json-store-handler.ts
@@ -1,7 +1,7 @@
 import { from, Observable, of } from 'rxjs';
-import { TimedTypedKeyValueStore } from '@digita-ai/handlersjs-storage';
+import { TimedTypedKeyValueStore } from '@useid/handlersjs-storage';
 import { map, switchMap } from 'rxjs/operators';
-import { getLogger } from '@digita-ai/handlersjs-logging';
+import { getLogger } from '@useid/handlersjs-logging';
 import { HttpHandler } from '../models/http-handler';
 import { HttpHandlerContext } from '../models/http-handler-context';
 import { HttpHandlerResponse } from '../models/http-handler-response';
@@ -26,9 +26,10 @@ export class JsonStoreHandler<T extends string, M extends { [t in T]: unknown }>
    *
    * @returns the stringified storage data as a HTTP response, or a Not Found HTTP response
    */
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   private tryProvideData(context: HttpHandlerContext): Observable<HttpHandlerResponse> {
 
-    return from(this.store.get(this.data)).pipe(map((data) =>  {
+    return from(this.store.get(this.data)).pipe(map((data) => {
 
       if (data) {
 
@@ -76,7 +77,7 @@ export class JsonStoreHandler<T extends string, M extends { [t in T]: unknown }>
 
           return of({ body: '', headers: {}, status: 304 });
 
-        }) // not modified
+        }), // not modified
       );
 
     } else {