+++ packages/handlersjs-http/lib/models/http-handler.ts
@@ -1,7 +1,7 @@
-import { Handler } from '@digita-ai/handlersjs-core';
+import { Handler } from '@useid/handlersjs-core';
 import { HttpHandlerResponse } from './http-handler-response';
 import { HttpHandlerContext } from './http-handler-context';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export abstract class HttpHandler<C extends HttpHandlerContext = HttpHandlerContext>
-  extends Handler<C, HttpHandlerResponse> {}
+  extends Handler<C, HttpHandlerResponse> { }