+++ packages/handlersjs-logging/lib/loggers/pino-logger.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-console -- this is a logger service */
-import { HandlerArgumentError } from '@digita-ai/handlersjs-core';
+import { HandlerArgumentError } from '@useid/handlersjs-core';
 import Pino from 'pino';
 import pretty from 'pino-pretty';
 import { LoggerLevel } from '../models/logger-level';
@@ -31,6 +31,7 @@ export class PinoLogger extends Logger {
 
   log(level: LoggerLevel, message: string, data?: unknown): void {
 
+    // eslint-disable-next-line no-null/no-null
     if (level === null || level === undefined) {
 
       throw new HandlerArgumentError('level should be set', level);